+++ src/functions/public/Expand-PermissionIdentity.ps1
@@ -1,0 +1,138 @@
+function Expand-PermissionIdentity {
+
+    param (
+
+        # Permission objects from Get-FolderAccessList whose IdentityReference to resolve
+        [Parameter(ValueFromPipeline)]
+        [object[]]$Identity,
+
+        # Output stream to send the log messages to
+        [ValidateSet('Silent', 'Quiet', 'Success', 'Debug', 'Verbose', 'Output', 'Host', 'Warning', 'Error', 'Information', $null)]
+        [string]$DebugOutputStream = 'Debug',
+
+        # Maximum number of concurrent threads to allow
+        [int]$ThreadCount = (Get-CimInstance -ClassName CIM_Processor | Measure-Object -Sum -Property NumberOfLogicalProcessors).Sum,
+
+        # Cache of known Win32_Account instances keyed by domain and SID
+        [hashtable]$Win32AccountsBySID = ([hashtable]::Synchronized(@{})),
+
+        # Cache of known Win32_Account instances keyed by domain (e.g. CONTOSO) and Caption (NTAccount name e.g. CONTOSO\User1)
+        [hashtable]$Win32AccountsByCaption = ([hashtable]::Synchronized(@{})),
+
+        <#
+        Dictionary to cache directory entries to avoid redundant lookups
+
+        Defaults to an empty thread-safe hashtable
+        #>
+        [hashtable]$DirectoryEntryCache = ([hashtable]::Synchronized(@{})),
+
+        # Hashtable with known domain NetBIOS names as keys and objects with Dns,NetBIOS,SID,DistinguishedName properties as values
+        [hashtable]$DomainsByNetbios = ([hashtable]::Synchronized(@{})),
+
+        # Hashtable with known domain SIDs as keys and objects with Dns,NetBIOS,SID,DistinguishedName properties as values
+        [hashtable]$DomainsBySid = ([hashtable]::Synchronized(@{})),
+
+        # Hashtable with known domain DNS names as keys and objects with Dns,NetBIOS,SID,DistinguishedName,AdsiProvider,Win32Accounts properties as values
+        [hashtable]$DomainsByFqdn = ([hashtable]::Synchronized(@{})),
+
+        <#
+        Hostname of the computer running this function.
+
+        Can be provided as a string to avoid calls to HOSTNAME.EXE
+        #>
+        [string]$ThisHostName = (HOSTNAME.EXE),
+
+        <#
+        FQDN of the computer running this function.
+
+        Can be provided as a string to avoid calls to HOSTNAME.EXE and [System.Net.Dns]::GetHostByName()
+        #>
+        [string]$ThisFqdn = ([System.Net.Dns]::GetHostByName((HOSTNAME.EXE)).HostName),
+
+        # Username to record in log messages (can be passed to Write-LogMsg as a parameter to avoid calling an external process)
+        [string]$WhoAmI = (whoami.EXE),
+
+        # Dictionary of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
+        [hashtable]$LogMsgCache = $Global:LogMessages,
+
+        <#
+        Do not get group members (only report the groups themselves)
+
+        Note: By default, the -ExcludeClass parameter will exclude groups from the report.
+          If using -NoGroupMembers, you most likely want to modify the value of -ExcludeClass.
+          Remove the 'group' class from ExcludeClass in order to see groups on the report.
+        #>
+        [switch]$NoGroupMembers
+
+    )
+
+    $LogParams = @{
+        LogMsgCache  = $LogMsgCache
+        ThisHostname = $ThisHostname
+        Type         = $DebugOutputStream
+        WhoAmI       = $WhoAmI
+    }
+
+    if ($ThreadCount -eq 1) {
+        $ExpandIdentityReferenceParams = @{
+            DirectoryEntryCache    = $DirectoryEntryCache
+            IdentityReferenceCache = $IdentityReferenceCache
+            DomainsBySID           = $DomainsBySID
+            DomainsByNetbios       = $DomainsByNetbios
+            DomainsByFqdn          = $DomainsByFqdn
+            ThisHostName           = $ThisHostName
+            ThisFqdn               = $ThisFqdn
+            WhoAmI                 = $WhoAmI
+            LogMsgCache            = $LogMsgCache
+        }
+        if ($NoGroupMembers) {
+            $ExpandIdentityReferenceParams['NoGroupMembers'] = $true
+        }
+
+        [int]$ProgressInterval = [math]::max(($Identity.Count / 100), 1)
+        $ProgressCounter = 0
+        $i = 0
+        ForEach ($ThisID in $Identity) {
+            $ProgressCounter++
+            if ($ProgressCounter -eq $ProgressInterval) {
+                $PercentComplete = $i / $Identity.Count * 100
+                Write-Progress -Activity 'Expand-IdentityReference' -Status "$([int]$PercentComplete)%" -CurrentOperation $ThisID.Name -PercentComplete $PercentComplete
+                $ProgressCounter = 0
+            }
+            $i++
+
+            $ExpandIdentityReferenceParams['AccessControlEntry'] = $ThisID
+            Write-LogMsg @LogParams -Text "Expand-IdentityReference -AccessControlEntry $($ThisID.Name)"
+            Expand-IdentityReference @ExpandIdentityReferenceParams
+        }
+        Write-Progress -Activity 'Expand-IdentityReference' -Completed
+
+    } else {
+        $ExpandIdentityReferenceParams = @{
+            Command              = 'Expand-IdentityReference'
+            InputObject          = $Identity
+            InputParameter       = 'AccessControlEntry'
+            TodaysHostname       = $ThisHostname
+            WhoAmI               = $WhoAmI
+            LogMsgCache          = $LogMsgCache
+            Threads              = $ThreadCount
+            AddParam             = @{
+                DirectoryEntryCache    = $DirectoryEntryCache
+                IdentityReferenceCache = $IdentityReferenceCache
+                DomainsBySID           = $DomainsBySID
+                DomainsByNetbios       = $DomainsByNetbios
+                DomainsByFqdn          = $DomainsByFqdn
+                ThisHostName           = $ThisHostName
+                ThisFqdn               = $ThisFqdn
+                WhoAmI                 = $WhoAmI
+                LogMsgCache            = $LogMsgCache
+            }
+            ObjectStringProperty = 'Name'
+        }
+        if ($NoGroupMembers) {
+            $ExpandIdentityReferenceParams['AddSwitch'] = 'NoGroupMembers'
+        }
+        Write-LogMsg @LogParams -Text "Split-Thread -Command 'Expand-IdentityReference' -InputParameter 'AccessControlEntry' -InputObject `$Identity"
+        Split-Thread @ExpandIdentityReferenceParams
+    }
+}