+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.86] - 2024-02-02 - added functionality from controller script export-permission.ps1
+
+## [0.0.85] - 2024-02-02 - added functionality from controller script export-permission.ps1
+
+## [0.0.84] - 2024-02-02 - added functionality from controller script export-permission.ps1
+
+## [0.0.83] - 2024-02-02 - added functionality from controller script export-permission.ps1
+
+## [0.0.82] - 2024-02-02 - added functionality from controller script export-permission.ps1
+
 ## [0.0.81] - 2024-01-31 - bugfix get-folderaccesslist
 
 ## [0.0.80] - 2024-01-31 - bugfix get-folderaccesslist